feat: 최단 경로 구하는 클래스 인터페이스로 추상화
