+++ src/main/java/subway/controller/section/SectionController.java
@@ -1,15 +1,12 @@
 package subway.controller.section;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.controller.section.dto.PathRequest;
 import subway.controller.section.dto.SectionInsertWebRequest;
 import subway.service.section.SectionService;
-import subway.service.section.dto.PathResult;
 import subway.service.section.dto.SectionCreateRequest;
 
 @RestController
@@ -31,11 +28,5 @@ public class SectionController {
                 sectionInsertWebRequest.getLineId());
         sectionService.insert(sectionCreateRequest);
         return ResponseEntity.ok().build();
-    }
-
-    @GetMapping
-    public ResponseEntity<PathResult> requestPath(@RequestBody PathRequest pathRequest) {
-        PathResult pathResult = sectionService.calculateShortestPathFee(pathRequest.getSourceStationId(), pathRequest.getTargetStationId());
-        return ResponseEntity.ok(pathResult);
     }
 }
