+++ src/main/java/subway/service/path/domain/FeePolicy.java
@@ -1,0 +1,7 @@
+package subway.service.path.domain;
+
+import java.util.List;
+
+public interface FeePolicy {
+    int calculateFee(List<SectionEdge> edges);
+}
