+++ src/main/java/subway/service/path/domain/JgraphtRoute.java
@@ -1,24 +1,21 @@
-package subway.service.section.domain;
+package subway.service.path.domain;
 
 import org.jgrapht.Graph;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.WeightedMultigraph;
+import org.springframework.stereotype.Component;
+import subway.service.path.dto.ShortestPath;
+import subway.service.section.domain.Section;
 import subway.service.station.domain.Station;
 
 import java.util.List;
 
-public class JgraphtRoute {
+@Component
+public class JgraphtRoute implements PathRouter {
 
-    private final List<Station> stations;
-    private final List<SectionEdge> edges;
-
-    private JgraphtRoute(List<Station> stations, List<SectionEdge> edges) {
-        this.stations = stations;
-        this.edges = edges;
-    }
-
-    public static JgraphtRoute of(List<Section> sections, Station source, Station target) {
+    @Override
+    public ShortestPath findShortestPath(List<Section> sections, Station source, Station target) {
         Graph<Station, SectionEdge> weightedGraph = new WeightedMultigraph<>(SectionEdge.class);
         for (Section section : sections) {
             Station upStation = section.getUpStation();
@@ -30,14 +27,6 @@ public class JgraphtRoute {
 
         DijkstraShortestPath<Station, SectionEdge> dijkstraShortestPath = new DijkstraShortestPath<>(weightedGraph);
         GraphPath<Station, SectionEdge> shortestPath = dijkstraShortestPath.getPath(source, target);
-        return new JgraphtRoute(shortestPath.getVertexList(), shortestPath.getEdgeList());
-    }
-
-    public List<Station> getStations() {
-        return stations;
-    }
-
-    public List<SectionEdge> getEdges() {
-        return edges;
+        return new ShortestPath(shortestPath.getVertexList(), shortestPath.getEdgeList());
     }
 }
