--- src/main/java/subway/service/section/domain/FeePolicy.java
@@ -1,7 +1,0 @@
-package subway.service.section.domain;
-
-import java.util.List;
-
-public interface FeePolicy {
-    public int calculateFee(List<SectionEdge> edges);
-}
